paradedb/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "plux>=1.3.1"]
build-backend = "setuptools.build_meta"

[project]
name = "localstack-extension-paradedb"
version = "0.1.0"
description = "LocalStack Extension: ParadeDB on LocalStack"
readme = {file = "README.md", content-type = "text/markdown; charset=UTF-8"}
requires-python = ">=3.10"
authors = [
    { name = "LocalStack team"}
]
keywords = ["LocalStack", "ParadeDB", "PostgreSQL"]
classifiers = []
dependencies = []

[project.urls]
Homepage = "https://github.com/localstack/localstack-extensions"

[project.optional-dependencies]
dev = [
    "boto3",
    "build",
    "jsonpatch",
    "localstack",
    "psycopg2-binary",
    "pytest",
    "rolo",
    "ruff",
]

[project.entry-points."localstack.extensions"]
localstack_paradedb = "localstack_paradedb.extension:ParadeDbExtension"
